move live classifiers into package
